Fix #1 (the way the media URL is constructed)
When accessing programs via "Toutes les émissions", the prefix is not part
of the href tag (as before), but when going directly to "@rrêt sur images",
the prefix is part of the href.
We check if the href starts with "http" or not.

Thanks to mossroy!

(cherry picked from commit 70d7b1f28912a96a5ddc3cdc2df2e67c75a3f0e0)